Use Dagger to create an ObjectMapper singleton and inject it into Czars.
