+++ build.gradle
@@ -10,6 +10,9 @@ plugins {
     // IntelliJ integration
     id 'idea'
 
+    // Dagger
+    id "net.ltgt.apt" version "0.10"
+
     // Apply the java plugin to add support for Java
     id 'java'
 
@@ -51,6 +54,10 @@ dependencies {
     compile "org.lwjgl:lwjgl-stb:${lwjglVersion}:natives-linux"
     compile "org.lwjgl:lwjgl-stb:${lwjglVersion}:natives-macos"
     compile "org.lwjgl:lwjgl-stb:${lwjglVersion}:natives-windows"
+
+    // Dagger
+    compile 'com.google.dagger:dagger:2.16'
+    apt 'com.google.dagger:dagger-compiler:2.16'
 
     // Use JUnit test framework
     testCompile 'junit:junit:4.12'
