+++ src/main/java/ure/terrain/UTerrainCzar.java
@@ -2,20 +2,12 @@ package ure.terrain;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Field;
 import java.util.HashMap;
-import java.util.Set;
 
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.ObjectCodec;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import org.reflections.Reflections;
-import org.reflections.scanners.SubTypesScanner;
+import ure.Injector;
+
+import javax.inject.Inject;
 
 /**
  * Load and manage all the terrain types and dole them out as needed.
@@ -30,11 +22,10 @@ public class UTerrainCzar {
     private  HashMap<Character,TerrainI> terrains;
     private  HashMap<String,TerrainI> terrainsByName;
 
-    private Set<Class<? extends TerrainI>> terrainClasses;
-    private ObjectMapper objectMapper = new ObjectMapper();
-    private Reflections reflections = new Reflections("ure", new SubTypesScanner());
-
     private Class<TerrainDeco> decorator;
+
+    @Inject
+    ObjectMapper objectMapper;
 
     /**
      * When creating your TerrainCzar, pass in your TerrainDeco class if you have one to extend
@@ -43,18 +34,18 @@ public class UTerrainCzar {
      * @param myDecorator
      */
     public UTerrainCzar(String jsonfilename, Class<TerrainDeco> myDecorator) {
-        SimpleModule module = new SimpleModule();
-        module.addDeserializer(TerrainI.class, new TerrainDeserializer());
-        objectMapper.registerModule(module);
+        this();
         decorator = myDecorator;
         loadTerrains(jsonfilename);
     }
 
     public UTerrainCzar(Class<TerrainDeco> myDecorator) {
-        SimpleModule module = new SimpleModule();
-        module.addDeserializer(TerrainI.class, new TerrainDeserializer());
-        objectMapper.registerModule(module);
+        this();
         decorator = myDecorator;
+    }
+
+    public UTerrainCzar() {
+        Injector.getAppComponent().inject(this);
     }
 
     /**
@@ -64,12 +55,6 @@ public class UTerrainCzar {
      *
      */
     public void loadTerrains(String resourceName) {
-        terrainClasses = reflections.getSubTypesOf(TerrainI.class);
-        System.out.println("+++++++++ " + terrainClasses.size() + " TerrainI subclasses");
-        for (Class<? extends TerrainI> terrainClass : terrainClasses) {
-            System.out.println("**** " + terrainClass.toString());
-        }
-        System.out.println("+++++++++");
         terrains = new HashMap<>();
         terrainsByName = new HashMap<>();
         try {
@@ -114,38 +99,4 @@ public class UTerrainCzar {
     public UTerrain getTerrainByName(String name) {
         return (UTerrain)(getTerrainForFilechar(terrainsByName.get(name).filechar));
     }
-
-    public class TerrainDeserializer extends JsonDeserializer<TerrainI> {
-
-        @Override
-        public TerrainI deserialize(JsonParser parser, DeserializationContext context) throws IOException, JsonProcessingException {
-            ObjectCodec codec = parser.getCodec();
-            JsonNode node = codec.readTree(parser);
-            JsonNode typeNode = node.get("type");
-            String type = typeNode != null ? node.get("type").asText() : null;
-            Class<? extends TerrainI> terrainClass = classForType(type);
-            return (objectMapper.treeToValue(node, terrainClass));
-        }
-
-        private Class<? extends TerrainI> classForType(String type) {
-            if (type == null || type.equals("")) {
-                return Blank.class;
-            }
-            try {
-                for (Class<? extends TerrainI> terrainClass : terrainClasses) {
-                    Field typeField = terrainClass.getField("TYPE");
-                    String typeValue = (String) typeField.get(null);
-                    if (type.equals(typeValue)) {
-                        return terrainClass;
-                    }
-                }
-                throw new RuntimeException("No TerrainI class of type '" + type + "' was found in the classpath");
-            } catch (NoSuchFieldException e) {
-                throw new RuntimeException("All subclasses of TerrainI must define a static TYPE field", e);
-            } catch (IllegalAccessException e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
 }
